% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdinaryLeastSquares.R
\name{fit_OLS}
\alias{fit_OLS}
\title{Ordinary Least Squares}
\usage{
fit_OLS(design_X, Y, penalty = "None")
}
\arguments{
\item{design_X}{}
}
\description{
This Package
}
\examples{
hello("World")

}
